requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.1
stix2>=3.0.1
stix2-validator>=3.2.0